gprofng: Fix -Wformat= warnings
Added format attribute to several gprofng functions.
Fixed -Wformat= warnings.

gprofng/ChangeLog
2023-10-23  Vladimir Mezentsev  <vladimir.mezentsev@oracle.com>

	* libcollector/heaptrace.c: Fixed -Wformat= warnings.
	* libcollector/hwprofile.c: Likewise.
	* libcollector/iolib.c: Likewise.
	* libcollector/iotrace.c: Likewise.
	* libcollector/jprofile.c: Likewise.
	* libcollector/profile.c: Likewise.
	* libcollector/synctrace.c: Likewise.
	* src/ClassFile.cc: Likewise.
	* src/SourceFile.cc: Likewise.
	* libcollector/libcol_util.h: Added format attribute.
	* src/Emsg.h: Likewise.
	* src/collector_module.h: Likewise.
	* src/data_pckts.h: Define fld_sizeof.

+++ gprofng/libcollector/jprofile.c
@@ -28,7 +28,7 @@
 #include <unistd.h>
 #include <errno.h>
 #include <sys/param.h> /* MAXPATHLEN */
-
+#include <stddef.h>
 #include <jni.h>
 #include <jvmti.h>
 
@@ -628,8 +628,8 @@ jvmti_ThreadStart (jvmtiEnv *jvmti_env, JNIEnv* jni_env, jthread thread)
 				 group_name,
 				 parent_name,
 				 (unsigned long) tid,
-				 thread,
-				 jni_env
+				 (unsigned long) thread,
+				 (unsigned long) jni_env
 				 );
   TSD_Entry *tsd = collector_interface->getKey (tsd_key);
   if (tsd)
@@ -648,8 +648,8 @@ jvmti_ThreadEnd (jvmtiEnv *jvmti_env, JNIEnv* jni_env, jthread thread)
 				 SP_JCMD_JTHREND,
 				 (unsigned) (hrt / NANOSEC), (unsigned) (hrt % NANOSEC),
 				 (unsigned long) tid,
-				 thread,
-				 jni_env
+				 (unsigned long) thread,
+				 (unsigned long) jni_env
 				 );
   TSD_Entry *tsd = collector_interface->getKey (tsd_key);
   if (tsd)
